// src/modules/excel-reports/domain/enums/ticket-report/ticket-report-estado.ts
export enum TicketReporteEstado {
  NUEVO = 'NUEVO',
  ABIERTA = 'ABIERTA',
  EN_PROCESO = 'EN_PROCESO',
  PENDIENTE = 'PENDIENTE',
  PENDIENTE_CLIENTE = 'PENDIENTE_CLIENTE',
  PENDIENTE_TECNICO = 'PENDIENTE_TECNICO',
  PENDIENTE_REVISION = 'PENDIENTE_REVISION',
  RESUELTA = 'RESUELTA',
  CERRADO = 'CERRADO',
  CANCELADA = 'CANCELADA',
  ARCHIVADA = 'ARCHIVADA',
}

// export const TicketReporteEstado = {
//   NUEVO: 'NUEVO',
//   ABIERTA: 'ABIERTA',
//   EN_PROCESO: 'EN_PROCESO',

//   PENDIENTE: 'PENDIENTE',
//   PENDIENTE_CLIENTE: 'PENDIENTE_CLIENTE',
//   PENDIENTE_TECNICO: 'PENDIENTE_TECNICO',
//   PENDIENTE_REVISION: 'PENDIENTE_REVISION',

//   RESUELTA: 'RESUELTA',
//   CERRADO: 'CERRADO',

//   CANCELADA: 'CANCELADA',
//   ARCHIVADA: 'ARCHIVADA',
// };

// export type TicketReporteEstado =
//   (typeof TicketReporteEstado)[keyof typeof TicketReporteEstado];