// src/modules/excel-reports/domain/enums/ticket-report/ticket-reporte-prioridad.enum.ts
export enum TicketReportePrioridad {
  BAJA = 'BAJA',
  MEDIA = 'MEDIA',
  ALTA = 'ALTA',
  URGENTE = 'URGENTE',
}

// export const TicketReportePrioridad = {
//   BAJA: 'BAJA',
//   MEDIA: 'MEDIA',
//   ALTA: 'ALTA',
//   URGENTE: 'URGENTE',
// } as const;

// export type TicketReportePrioridad =
//   (typeof TicketReportePrioridad)[keyof typeof TicketReportePrioridad];